+++ src/test/java/org/sbrubbles/conditio/fixtures/tracing/WorkDone.java
@@ -2,7 +2,7 @@ package org.sbrubbles.conditio.fixtures.tracing;
 
 import org.sbrubbles.conditio.Condition;
 
-public class WorkDone extends Condition {
+public class WorkDone implements Condition {
   private final int amount;
 
   public WorkDone(int amount) {